pyproject.toml
[project]
name = "data-science-ghcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "joblib>=1.5.3",
    "mlflow-skinny>=2.20",
    "optuna>=4.8.0",
    "pyarrow>=24.0.0",
    "pydantic>=2.7",
    "scikit-learn>=1.8.0",
    "shap>=0.51.0",
    "uvicorn[standard]>=0.30",
    "xgboost>=3.2.0",
]
